requirements.txt
python-telegram-bot==21.6
aiohttp==3.9.5
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
openai==1.40.0